requirements.txt
google-api-python-client
six==1.9.0
requests>=2.4.0
pyVimeo
oauth2client
